refactor: convert all Dashboard styles to contextualized SASS
