+++ components/dashboard/Dashboard.module.scss
@@ -1,3 +1,5 @@
+@import "../../styles/variables";
+
 /**
 DASHBOARD
 ===========================
@@ -12,6 +14,9 @@ DASHBOARD
     width: 90%;
   }
 }
+
+
+// Balance
 
 .dashboard-header {
   display: flex;
@@ -117,7 +122,7 @@ DASHBOARD
   }
   
   button {
-    color: #fff;
+    color: $white;
     border: 1px solid $green;
     background: rgba(255, 255, 255, 0.01);
     margin-left: 7px;
@@ -125,6 +130,9 @@ DASHBOARD
     height: 25px;
   }
 }
+
+
+// Home calls
 
 .home-calls {
   display: flex;
@@ -221,3 +229,90 @@ DASHBOARD
 .disabled {
   opacity: .35;
 }
+
+
+// Payments
+
+.payments-section {
+  padding-bottom: 100px;
+
+  @media #{$smartphones-1} {
+    padding-bottom: 50px;
+  }
+
+  h2 {
+    margin: 0 0 30px 0;
+    font-size: $font-large-2;
+    font-weight: 600;
+
+    @media #{$smartphones-1} {
+      font-size: $font-large;
+    }
+  }
+}
+
+.payment-item {
+  margin-bottom: 20px;
+  width: 100%;
+  border: 1px solid $dark-blue-2;
+  background: rgba($medium-blue, .07);
+  border-radius: 6px;
+  font-size: 14px;
+  color: $white;
+  box-sizing: border-box;
+  display: block;
+  margin: 8px 0 15px 0;
+  padding: 20px;
+  @extend %clearfix;
+  position: relative;
+
+  h3 {
+    font-size: $font-medium-2;
+    padding: 0 0 10px 0;
+    font-weight: 600;
+  }
+
+  a {
+    color: $medium-gray;
+    font-size: $font-small;
+    text-decoration: none;
+  }
+
+  .value-status {
+    position: absolute;
+    text-align: right;
+    right: 20px;
+    top: 20px;
+
+    em {
+      font-size: $font-large;
+      display: block;
+      font-style: normal;
+      margin-bottom: 5px;
+    }
+
+    img {
+      width: 20px;
+      float: right;
+      margin: -2px 0 0 5px;
+    }
+
+    span {
+      display: inline-block;
+      font-size: 9px;
+    }
+
+    strong {
+      font-weight: normal;
+    }
+
+    .paid {
+      color: $medium-gray;
+    }
+
+    .received {
+      color: $green;
+    }
+  }
+}
+
