+++ styles/globals.scss
@@ -13,8 +13,6 @@ GLOBAL
 @import "forms";
 @import "footer";
 @import "connect-wallet";
-@import "dashboard";
-@import "payments";
 @import "components";
